Create a Bootstrap 386 tab panel

Create a Bootstrap 386 tab panel

Usage

tab_panel_386(title, ..., value = title, icon = NULL)

Arguments

title

Display title for tab

...

UI elements to include within the tab

value

The value that should be sent when tabsetPanel reports that this tab is selected. If omitted and tabsetPanel has an id, then the title will be used.

icon

Optional icon to appear on the tab. This attribute is only valid when using a tabPanel within a navbarPage().